Moved controllers to their own folder, and stores to lib rather than shared
