+++ ui/src/GamezPane/GamezPane.svelte
@@ -7,11 +7,11 @@
   import { decodeHashFromBase64 } from '@holochain/client';
   import { isWeContext, weaveUrlFromWal, type WAL } from '@lightningrodlabs/we-applet';
 
-  import { type BoardState, PieceDef, Board, type Piece, type BoardProps } from '~/shared/board';
-  import { type AssetSpec } from '~/shared/util';
+  import { type BoardState, PieceDef, Board, type Piece, type BoardProps } from '~/lib/board';
+  import { getStoreContext } from '~/lib/context';
+  import { type AssetSpec } from '~/lib/util';
   import EditBoardDialog from '~/shared/EditBoardDialog.svelte';
   import PlayerName from '~/shared/PlayerName.svelte';
-  import { getStoreContext } from '~/lib/context';
 
   import PieceAttachmentDialog from './PieceAttachmentsDialog.svelte';
   import WalSpace from './WalSpace.svelte';
