+++ ui/src/lib/context.ts
@@ -1,6 +1,5 @@
 import { getContext, setContext } from 'svelte';
-
-import { GamezStore } from '../shared/store';
+import { GamezStore } from '~/lib/store';
 
 export const setStoreContext = (getter: () => GamezStore) => setContext('store', { store: getter });
 export const getStoreContext = () => getContext<{ store: () => GamezStore }>('store').store();
